+++ core/src/main/scala/com/krux/hyperion/h3/activity/PipelineActivity.scala
@@ -17,58 +17,57 @@ trait PipelineActivity[A <: ResourceObject] extends PipelineObject {
 
   type Self <: PipelineActivity[A]
 
-  def activityFields: ActivityFields[A]
   def activityFieldsLens: Lens[Self, ActivityFields[A]]
 
-  def dependsOn = activityFields.dependsOn
+  def dependsOn = (activityFieldsLens >> 'dependsOn).get(self)
   private[hyperion] def dependsOn(activities: PipelineActivity[_]*): Self =
     (activityFieldsLens >> 'dependsOn).modify(self)(_ ++ activities)
 
-  def preconditions = activityFields.preconditions
+  def preconditions = (activityFieldsLens >> 'preconditions).get(self)
   def whenMet(conditions: Precondition*): Self =
     (activityFieldsLens >> 'preconditions).modify(self)(_ ++ conditions)
 
-  def onFailAlarms = activityFields.onFailAlarms
+  def onFailAlarms = (activityFieldsLens >> 'onFailAlarms).get(self)
   def onFail(alarms: SnsAlarm*): Self =
     (activityFieldsLens >> 'onFailAlarms).modify(self)(_ ++ alarms)
 
-  def onSuccessAlarms = activityFields.onSuccessAlarms
+  def onSuccessAlarms = (activityFieldsLens >> 'onSuccessAlarms).get(self)
   def onSuccess(alarms: SnsAlarm*): Self =
     (activityFieldsLens >> 'onSuccessAlarms).modify(self)(_ ++ alarms)
 
-  def onLateActionAlarms = activityFields.onLateActionAlarms
+  def onLateActionAlarms = (activityFieldsLens >> 'onLateActionAlarms).get(self)
   def onLateAction(alarms: SnsAlarm*): Self =
     (activityFieldsLens >> 'onLateActionAlarms).modify(self)(_ ++ alarms)
 
-  def maximumRetries = activityFields.maximumRetries
+  def maximumRetries = (activityFieldsLens >> 'maximumRetries).get(self)
   def withMaximumRetries(retries: HInt): Self =
     (activityFieldsLens >> 'maximumRetries).set(self)(Option(retries))
 
-  def attemptTimeout = activityFields.attemptTimeout
+  def attemptTimeout = (activityFieldsLens >> 'attemptTimeout).get(self)
   def withAttemptTimeout(duration: HDuration): Self =
     (activityFieldsLens >> 'attemptTimeout).set(self)(Option(duration))
 
-  def lateAfterTimeout = activityFields.lateAfterTimeout
+  def lateAfterTimeout = (activityFieldsLens >> 'lateAfterTimeout).get(self)
   def withLateAfterTimeout(duration: HDuration): Self =
     (activityFieldsLens >> 'lateAfterTimeout).set(self)(Option(duration))
 
-  def retryDelay = activityFields.retryDelay
+  def retryDelay = (activityFieldsLens >> 'retryDelay).get(self)
   def withRetryDelay(duration: HDuration): Self =
     (activityFieldsLens >> 'retryDelay).set(self)(Option(duration))
 
-  def failureAndRerunMode = activityFields.failureAndRerunMode
+  def failureAndRerunMode = (activityFieldsLens >> 'failureAndRerunMode).get(self)
   def withFailureAndRerunMode(mode: FailureAndRerunMode): Self =
     (activityFieldsLens >> 'failureAndRerunMode).set(self)(Option(mode))
 
   // TODO: Uncomment the following once the other activities has been transformed
-  def objects: Iterable[PipelineObject] = activityFields.dependsOn // ++
+  def objects: Iterable[PipelineObject] = dependsOn // ++
     // activityFields.onFailAlarms ++
     // activityFields.onSuccessAlarms ++
     // activityFields.onLateActionAlarms ++
     // activityFields.preconditions :+
     // activityFields.runsOn
 
-  def runsOn: Resource[A] = activityFields.runsOn
+  def runsOn: Resource[A] = (activityFieldsLens >> 'runsOn).get(self)
 
   def serialize: AdpActivity
   def ref: AdpRef[AdpActivity] = AdpRef(serialize)