+++ core/src/main/scala/com/krux/hyperion/h3/dataformat/DataFormatFields.scala
@@ -1,0 +1,5 @@
+package com.krux.hyperion.h3.dataformat
+
+import com.krux.hyperion.adt.HString
+
+case class DataFormatFields(columns: Seq[HString] = Seq.empty)