+++ core/src/main/scala/com/krux/hyperion/h3/datanode/Copyable.scala
@@ -1,0 +1,6 @@
+package com.krux.hyperion.h3.datanode
+
+/**
+ * The data node that can be used as input and output in CopyActivity
+ */
+trait Copyable extends DataNode